requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
scikit-learn
torch
tensorflow
keras
transformers
nltk
sentence-transformers
tensorboard
fastapi
uvicorn
flask
gradio
streamlit
pyarrow
h5py
optuna
hyperopt
plotly
tqdm
joblib
pytest
pydantic
langchain
langchain-core
langchain-openai
langfuse
huggingface-hub
ipykernel
jupyterlab
tavily-python